final project restructure
